chore(proj1a): Temporarily disable copy constructors to pass API checks
The autograder for this submission phase unexpectedly requires the copy constructors to be removed, despite being part of the project spec.

To pass the current API check, the copy constructors for both ArrayDeque and LinkedListDeque have been commented out. They are intended to be re-enabled for later project phases.

Also resolves remaining minor Checkstyle violations.